Initial stas-sdk code framework (ALPHA)
 The code includes various files and functions for utilizing STAS tokens (see stastoken.com for more details)

It's worth noting that the code is currently in the ALPHA phase, which means that it is still in the early stages of development and may contain bugs or other issues. Therefore, it's important to thoroughly test the code before using it in a production environment.

During the ALPHA phase, the development team will be working to improve the code and address any issues that arise. Public Pull Requests are also welcome. As the project progresses, additional features and functionality will be added to the codebase, and the project will move into subsequent phases of development.

Overall, this addition of code to the stas-sdk is an exciting step forward, and we are looking forward to seeing the library evolve and mature over time with input and feedback from the community.