op code conversion
Added in op code conversion function to data formatting.

This will allow op code string values to to be added to the data array  for issuance or transfer.

Op code constant values added.

If any strings match one of the key of the op code object it will then return its value.

Unit Test cases added in for new functions