Return un-serialized transactions
As ARC broadcast endpoints are poised to become the new standard, we're updating our library to make the conversion to the extended format seamless, negating the necessity of handling additional input data for this process.

For optimal usage of ARC, the transaction must be converted to the extended format, which requires the scriptPubKey and satoshi value for each input. By ensuring the non-serialized transaction is returned, these components will already be included within the transaction returned by the functions. This eliminates the need for any subsequent intervention to reintroduce this data before transitioning to the extended format.

For more information on the ARC transaction processor please see this documentation
https://bitcoin-sv.github.io/arc/#/

To continue using the library as per normal it will require that transactions are serialised once returned from the functions.

In addition a bug fix has been added for a utility function regarding BN.js library